commit final

+++ project-front/src/components/CardComponentView.tsx
@@ -1,4 +1,5 @@
 import styles from './CardView.module.css';
+import { format } from 'date-fns';
 import { AuthContext } from '../contexts/auth';
 import { useContext, useState} from 'react';
 import CardComponentEdit from './CardComponentEdit';
@@ -8,45 +9,53 @@ type CardComponentViewProps = {
   content: string;
   title: string;
   id: number;
+  date: string;
 };
 
-const CardComponentView: React.FC<CardComponentViewProps> = ({ id, content, title }) => {
+const CardComponentView: React.FC<CardComponentViewProps> = ({ id, content, title, date }) => {
   const { deleteNote } = useContext(AuthContext);
-  const [isModalOpen, setIsModalOpen] = useState(false);
+  const [isModalOpenEdit, setIsModalOpenEdit] = useState(false);
 
-  const openModal = () => {
-    setIsModalOpen(true);
+  const formatDate = (dateString:string) => {
+    const date = new Date(dateString);
+    return format(date, 'dd/MM/yyyy');
   };
 
-  const closeModal = () => {
-    setIsModalOpen(false);
+  const openModalEdit = () => {
+    setIsModalOpenEdit(true);
+  };
+
+  const closeModalEdit = () => {
+    setIsModalOpenEdit(false);
   };
 
   const handleDelete = () => {
-    window.location.reload();
-    deleteNote(id);
+    const response = window.confirm('Tem certeza que deseja excluir esta nota?')
+    if (response === true) {
+      deleteNote(id);
+      window.location.reload();
+    }
   };
 
   return (
     <>
-      <div className={styles.card}>
-          <a href="/view" className={styles.see} />
-          <a className={styles.edit} onClick={openModal}/>
-          <a className={styles.del} onClick={handleDelete} />
-          <h5 className={styles.cardtitle}>{title}</h5>
+      <div className={styles.card}>        
+        <a className={styles.edit} onClick={openModalEdit}/>
+        <a className={styles.del} onClick={handleDelete} />
+        <h5 className={styles.cardtitle}>{title}</h5>
         <div className={styles.cardbody}>
           <p className={styles.cardtext}>{content}</p>
         </div>
+        <p className={styles.date}>nota criada no dia: {formatDate(date)}</p>
       </div>
-
       <Modal
-        isOpen={isModalOpen}
-        onRequestClose={closeModal}
+        isOpen={isModalOpenEdit}
+        onRequestClose={closeModalEdit}
         contentLabel="Página de Registro"
         shouldCloseOnEsc={true}
         className={styles.modal}
       >
-        <button className={styles.close} onClick={closeModal}>X</button>
+        <button className={styles.close} onClick={closeModalEdit}>X</button>
         <CardComponentEdit key={id} id={id} initial_title={title} initial_content={content}/>
       </Modal>
     </>
